updating kop.sh md5 checking

+++ lib/kop.sh
@@ -10,18 +10,20 @@ varko() {
 	if [ -d $1 ] && [ $ != "" ]; then
 		debug KOP $1 $2
 		KR_LATEST_TODAY="today_maybe"
-		# collect md5 checksums of old and new
-		KR_MD5_OLD=$(md5sum $KR_DIR_BUT/$2.$USER.$HOSTNAME.old)
-		KR_MD5_NEW=$(md5sum $KR_DIR_BUT/$2.$USER.$HOSTNAME.zip)
-		echo $KR_MD5_OLD vs $KR_MD5_NEW
 		# add a comment to backup.log
 		bulog $1 $2.zip original_dir
 		debug KOP $KR_LATEST_TODAY
 		# doesn't do backup runs if already backed up today (checkup happens in bulog)
 		if [ $KR_LATEST_TODAY == "today_false" ]; then
-			rm $KR_DIR_BUT/$2.$USER.$HOSTNAME.old
-			mv $KR_DIR_BUT/$2.$USER.$HOSTNAME.zip $KR_DIR_BUT/$2.$USER.$HOSTNAME.old
-			zip -qr $KR_DIR_BUT/$2.$USER.$HOSTNAME.zip $1
+			if [ -f $KR_DIR_BUT/$2.$USER.$HOSTNAME.old ]; then rm $KR_DIR_BUT/$2.$USER.$HOSTNAME.old; fi
+			# collect md5 checksums of old backup
+			KR_MD5_OLD=$(md5sum $KR_DIR_BUT/$2.$USER.$HOSTNAME.zip)
+			if [ -f $KR_DIR_BUT/$2.$USER.$HOSTNAME.zip ]; then mv $KR_DIR_BUT/$2.$USER.$HOSTNAME.zip $KR_DIR_BUT/$2.$USER.$HOSTNAME.old; fi
+			zip -qr $KR_DIR_BUT/$2.$USER.$HOSTNAME.zip $1 -x "backup.log"
+			# collect md5 checksums of new and comparing it to old
+			KR_MD5_NEW=$(md5sum $KR_DIR_BUT/$2.$USER.$HOSTNAME.zip)
+			echo $KR_MD5_OLD vs $KR_MD5_NEW
+			# encrypting the backup file
 			gpg --encrypt -a -r $RECIPIENT $KR_DIR_BUT/$2.$USER.$HOSTNAME.zip
 			mv $KR_DIR_BUT/$2.$USER.$HOSTNAME.zip.asc $KR_DIR_BUC/$2.$USER.$HOSTNAME.cbc
 		else
